updated urls, tests and profile serializers
