+++ hackathon_site/event/api_urls.py
@@ -13,13 +13,13 @@ urlpatterns = [
         name="join-team",
     ),
     path("teams/leave_team/", api_views.LeaveTeamView.as_view(), name="leave-team"),
-    re_path(
-        r"^profiles/(?P<pk>[0-9]+)/$",
-        views.ProfileDetailView.as_view(),
+    path(
+        "profiles/<int:pk>/",
+        api_views.ProfileDetailView.as_view(),
         name="profile-detail",
     ),
     path(
-        "profiles/profile",
+        "profiles/profile/",
         api_views.CurrentProfileView.as_view(),
         name="current-profile",
     ),
