+++ hackathon_site/event/views.py
@@ -2,7 +2,6 @@ from datetime import datetime, timedelta
 
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.db import transaction
-from django.http import HttpResponseForbidden
 from django.shortcuts import redirect
 from django.urls import reverse_lazy
 from django.views.generic.base import TemplateView
@@ -12,19 +11,18 @@ from django.views.generic.edit import FormView
 from django.conf import settings
 from django_filters import rest_framework as filters
 
-from rest_framework import generics, mixins, status, permissions
+from rest_framework import generics, mixins
 from rest_framework.filters import SearchFilter
-from rest_framework.response import Response
 
 from hackathon_site.utils import is_registration_open
 from registration.forms import JoinTeamForm
 from registration.models import Team as RegistrationTeam
 
 
-from event.models import Team as EventTeam, Profile
-from event.serializers import TeamSerializer, ProfileModifySerializer, ProfileSerializer
+from event.models import Team as EventTeam
+from event.serializers import TeamSerializer
 from event.api_filters import TeamFilter
-from event.permissions import FullDjangoModelPermissions, UserHasProfile
+from event.permissions import FullDjangoModelPermissions
 
 
 def _now():
@@ -227,28 +225,3 @@ class TeamListView(mixins.ListModelMixin, generics.GenericAPIView):
 
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
-
-
-class ProfileDetailView(mixins.UpdateModelMixin, generics.GenericAPIView):
-
-    queryset = Profile.objects.all()
-    serializer_class = ProfileModifySerializer
-    permission_classes = [FullDjangoModelPermissions]
-
-    @transaction.atomic
-    def patch(self, request, *args, **kwargs):
-        profile = self.get_object()
-
-        data = request.data
-        profile.id_provided = data["id_provided"]
-        profile.attended = data["attended"]
-
-        if not profile.acknowledge_rules:
-            profile.acknowledge_rules = data["acknowledge_rules"]
-        if not profile.e_signature:
-            profile.e_signature = data["e_signature"]
-
-        profile.save()
-        response_serializer = ProfileSerializer(profile)
-        response_data = response_serializer.data
-        return Response(data=response_data, status=status.HTTP_200_OK,)
